replace navigation-path-script with props getter function
