+++ src/views/collections/collection-chapter.astro
@@ -1,5 +1,6 @@
 ---
 import styles from "./collection-chapter.module.scss";
+import { getHrefContainerProps } from "utils/href-container-script";
 
 interface CollectionChapter {
 	num: number;
@@ -11,7 +12,7 @@ interface CollectionChapter {
 const { num, title, description, href } = Astro.props as CollectionChapter;
 ---
 
-<div class={styles.container} data-navigation-path={href}>
+<div class={styles.container} {...getHrefContainerProps(href)}>
 	<div class={styles.numAndTitleContainer}>
 		<div class={styles.numContainer}>
 			<p class={`text-style-headline-6 ${styles.num}`} aria-hidden="true">
