+++ src/views/explore/homepage.astro
@@ -33,11 +33,6 @@ const unicornProfilePicMap = await getUnicornProfilePicMap();
 	/>
 </section>
 
-<script>
-	import { setupNavigationPaths } from "../../utils/navigation-path-script";
-	setupNavigationPaths();
-</script>
-
 <section>
 	<SubHeader tag="h1" text="Collections">
 		<Button href="/page/1">{translate(Astro, "action.view_all")}</Button>
