+++ src/views/explore/page.astro
@@ -21,9 +21,4 @@ const unicornProfilePicMap = await getUnicornProfilePicMap();
 	unicornProfilePicMap={unicornProfilePicMap}
 />
 
-<script>
-	import { setupNavigationPaths } from "../../utils/navigation-path-script";
-	setupNavigationPaths();
-</script>
-
 <Pagination page={page} client:load />
