+++ src/views/unicorn/unicorn-page.astro
@@ -25,8 +25,3 @@ const unicornProfilePicMap = await getUnicornProfilePicMap();
 		unicornProfilePicMap={unicornProfilePicMap}
 	/>
 </UnicornLayout>
-
-<script>
-	import { setupNavigationPaths } from "../../utils/navigation-path-script";
-	setupNavigationPaths();
-</script>
